refactor: share catalogue manifest rehydrate across servers
